requirements.txt
altair==5.2.0
anyio==4.2.0
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
boto3==1.34.17
botocore==1.34.17
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
extra-streamlit-components==0.1.60
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mysql-connector-python==8.2.0
numpy==1.26.3
packaging==23.2
pandas==2.1.4
pillow==10.2.0
protobuf==4.21.12
pyarrow==14.0.2
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
PyMySQL==1.1.0
pyTelegramBotAPI==4.14.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram==0.18.0
python-telegram-bot==20.7
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
rich==13.7.0
rpds-py==0.16.2
s3transfer==0.10.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.25
streamlit==1.30.0
streamlit-authenticator==0.2.3
telegram-text==0.1.2
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
watchdog==3.0.0
zipp==3.17.0
